Updated to 1.9.0
A new zip task has been created to create an archive with a project, added: checking and deleting an already created zip archive, error handling, creating a zip archive with a finished project for production.

A new ftp task has been created to work with a remote server, added: logging the process of sending files to the server, creating a connection to a remote server, error handling, sending the finished project to a folder on the remote server.

The deployZIP mode has been created, now when you run the npm run zip command, the process of creating and packaging the project into a zip archive for production will start.

The deployFTP mode has been created; now, when you run the npm run deploy command, the process of sending the finished project to a remote server will start.